refactor: movie as prop of Movie Card
