+++ src/pages/Search.tsx
@@ -48,25 +48,10 @@ export const MovieSearch = () => {
       ))}
       <Wrap spacing="1.875rem">
         {movieResultsList.map((movie, i) => {
-          const movieGenres = movie.genre_ids.map((genreId) => {
-            const genre = genres.find((genre) => genre.id === genreId);
-            return genre ? genre.name : '';
-          });
-          const genreNames = movieGenres.join(', ');
-          const releaseYear = movie.release_date.split('-')[0];
-          const rating = Number((movie.vote_average * 10).toFixed(0));
-
           return (
             <div key={i}>
               <WrapItem>
-                <MovieCard
-                  title={movie.title}
-                  overview={movie.overview}
-                  imageUrl={`https://image.tmdb.org/t/p/w400${movie.poster_path}`}
-                  release={releaseYear}
-                  genres={genreNames}
-                  rating={rating}
-                />
+                <MovieCard movie={movie} />
               </WrapItem>
             </div>
           );
